just some additional work on the DataFrame Store class

+++ src/sageworks/api/df_store.py
@@ -54,7 +54,7 @@ class DFStore:
         try:
             df = self.details()
             formatted_list = [
-                f"{row['name']} ({row['size'] / (1024 * 1024):.3f} MB) Last Modified: {row['modified'].strftime('%Y-%m-%d %H:%M:%S')}"
+                f"{row['name']}  ({row['size'] / (1024 * 1024):.3f}MB/{row['modified'].strftime('%Y-%m-%d %H:%M:%S')})"
                 for _, row in df.iterrows()
             ]
             return formatted_list
@@ -66,27 +66,28 @@ class DFStore:
         """Return a DataFrame with detailed metadata for all objects in the data_store prefix."""
         try:
             response = self.s3_client.list_objects_v2(Bucket=self.sageworks_bucket, Prefix=self.prefix)
-
-            if 'Contents' not in response:
-                return pd.DataFrame(columns=['name', 's3_file', 'size', 'created', 'modified'])
+            if "Contents" not in response:
+                return pd.DataFrame(columns=["name", "s3_file", "size", "modified"])
 
             # Collect details for each object
             data = []
-            for obj in response['Contents']:
-                name = obj["Key"].replace(f"{self.prefix}", "")
-                s3_file = f"s3://{self.sageworks_bucket}/{obj['Key']}"
+            for obj in response["Contents"]:
+                full_key = obj["Key"]
+
+                # Reverse logic: Strip the bucket/prefix in the front and .parquet in the end
+                name = full_key.replace(f"{self.prefix}", "").split(".parquet")[0]
+                s3_file = f"s3://{self.sageworks_bucket}/{full_key}"
                 size = obj["Size"]
-                created = obj["LastModified"]
                 modified = obj["LastModified"]
-                data.append([name, s3_file, size, created, modified])
+                data.append([name, s3_file, size, modified])
 
             # Create and return DataFrame
-            df = pd.DataFrame(data, columns=['name', 's3_file', 'size', 'created', 'modified'])
+            df = pd.DataFrame(data, columns=["name", "s3_file", "size", "modified"])
             return df
 
         except Exception as e:
             self.log.error(f"Failed to get object details: {e}")
-            return pd.DataFrame(columns=['name', 's3_file', 'size', 'created', 'modified'])
+            return pd.DataFrame(columns=["name", "s3_file", "size", "created", "modified"])
 
     def get(self, name: str) -> pd.DataFrame:
         """Retrieve a DataFrame from the AWS S3.
@@ -101,20 +102,16 @@ class DFStore:
         try:
             df = wr.s3.read_parquet(s3_uri)
             return df
-        except ClientError as e:
+        except ClientError:
             self.log.warning(f"Data '{name}' not found in S3.")
             return pd.DataFrame()  # Return an empty DataFrame if not found
-        except Exception as e:
-            self.log.error(f"Failed to retrieve data '{name}': {e}")
-            return pd.DataFrame()
 
-    def add(self, name: str, data: Union[pd.DataFrame, pd.Series], overwrite: bool = True):
+    def add(self, name: str, data: Union[pd.DataFrame, pd.Series]):
         """Add or update a DataFrame or Series in the AWS S3.
 
         Args:
             name (str): The name of the data.
             data (Union[pd.DataFrame, pd.Series]): The data to be stored.
-            overwrite (bool): Whether to overwrite existing data.
         """
         # Check if the data is a Pandas Series, convert it to a DataFrame
         if isinstance(data, pd.Series):
@@ -126,17 +123,11 @@ class DFStore:
 
         s3_uri = self._generate_s3_uri(name)
         try:
-            # Use 'overwrite' mode if overwrite is True, otherwise check for existence and raise an error
-            mode = 'overwrite' if overwrite else None
-            wr.s3.to_parquet(df=data, path=s3_uri, dataset=True, mode=mode)
+            wr.s3.to_parquet(df=data, path=s3_uri, dataset=True, mode="overwrite")
             self.log.info(f"Data '{name}' added/updated successfully in S3.")
         except Exception as e:
-            if not overwrite and wr.s3.does_object_exist(s3_uri):
-                self.log.error(f"Data '{name}' already exists in S3 and overwrite is set to False.")
-                raise FileExistsError(f"Data '{name}' already exists in S3.")
-            else:
-                self.log.critical(f"Failed to add/update data '{name}': {e}")
-                raise
+            self.log.critical(f"Failed to add/update data '{name}': {e}")
+            raise
 
     def delete(self, name: str):
         """Delete a DataFrame from the AWS S3.
@@ -145,6 +136,13 @@ class DFStore:
             name (str): The name of the data to delete.
         """
         s3_uri = self._generate_s3_uri(name)
+
+        # Check if the folder (prefix) exists in S3
+        if not wr.s3.list_objects(s3_uri):
+            self.log.warning(f"Data '{name}' does not exist in S3. Cannot delete.")
+            return
+
+        # Delete the data from S3
         try:
             wr.s3.delete_objects(s3_uri)
             self.log.info(f"Data '{name}' deleted successfully from S3.")
@@ -175,16 +173,20 @@ if __name__ == "__main__":
 
     # Add a new DataFrame
     df = pd.DataFrame({"A": [1, 2], "B": [3, 4]})
-    df_store.add("test_data", df, overwrite=True)
+    df_store.add("test_data", df)
 
     # Get the DataFrame
     print(f"Getting data 'test_data':\n{df_store.get('test_data')}")
 
-    # Delete the data
-    df_store.delete("test_data")
-
     # Now let's test adding a Series
     series = pd.Series([1, 2, 3, 4], name="Series")
-    df_store.add("test_series", series, overwrite=True)
+    df_store.add("test_series", series)
     print(f"Getting data 'test_series':\n{df_store.get('test_series')}")
+
+    # List the data again
+    print("Listing Data...")
+    print(df_store.list())
+
+    # Now delete the test data
+    df_store.delete("test_data")
     df_store.delete("test_series")
